Add ability types
This required adding some validation library (I used typia, but most
people use zod) to validate props. Here is a long term issue: 4 months
ago unplugin-typia was deprecated and archived by its owner, and nobody
stepped up to continue maintenance (not even personal forks in the
timeline):

https://github.com/ryoppippi/unplugin-typia

For now it works perfectly, but some day it might not.

Anyway I renamed the component to `AbilityNode` because for validation
it now defines a more abstract `Ability` inside it (an `AbilityNode` is
just the presentation of whatever data is defined on `Ability`).

There is also a stray button for toggling ability types but that was
just to test abilities are reactive (and also to save this styling
somewhere because I will need buttons), it will probably get removed.